revert: reverted refactor of types to avoid conflicts with victor's changes
